refactoring separate dto and service made
