+++ src/integrations/redis/dto/overallProgress.dto.ts
@@ -1,0 +1,13 @@
+import { ObjectType, Field } from '@nestjs/graphql';
+
+@ObjectType()
+export class OverallProgress {
+  @Field()
+  totalEmails: number;
+
+  @Field()
+  emailsProcessed: number;
+
+  @Field()
+  overallProgressPercentage: number;
+}
